Updated floats to doubles to reduce confusion. This increases the memory used, but not that much is being used in the first place. This is subject to change if memory becomes an issue in the future.

+++ src/ocean_models/include/fvcom/FVCOM.h
@@ -97,7 +97,7 @@ private:
 	 */
 	const FVCOMChunk::TriangleData& getTriangleData(int triangle, int siglayTriangleIndex, int timeIndex);
 
-	ModelData interpolate(FVCOMStructure::Point p, float time);
+	ModelData interpolate(FVCOMStructure::Point p, double time);
 	FVCOMChunk::NodeData nodeInterpolation(const FVCOMStructure::Point& interpolatedPoint, int containingTriangle, int siglayIndex, int timeIndex);
 	FVCOMChunk::TriangleData triangleInterpolation(const FVCOMStructure::Point& interpolatedPoint, int containingTriangle, int siglayIndex, int timeIndex);
 	const double areaOfTriangle(const FVCOMStructure::Point& p1, const FVCOMStructure::Point& p2, const FVCOMStructure::Point& p3) const;
